Omni frameworks as of OmniOutliner for iPad 1.0 submission (r148940):
General:
  - Fixed warnings from newer clang builds, mainly about not assigning to self
    in init.

OmniBase:
  - Add OBNSObjectCopyShortDescription() for CF callbacks (moved from
    OmniFoundation).
  - Whitelist some more system framework/classes from our runtime method
    checks.
  - Remove hack for STAssertThrows since it works in the latest Xcode 4.0.x
    builds.
  - Made OBAnalyzerNotReached() public. Useful for telling clang-sa about
    unreachable conditions.

OmniFoundation:
  - Speed up key searches in OFKnownKeyDictionary.
  - Add OFExtentIntersection().
  - Add some more support to OFFileWrapper.
    - Handle symlinks
    - Handle preferred file name conflicts.
    - Add serialization support (not byte compatible with NSFileWrapper,
      just API compatible).
  - Fixed some edge cases and bugs in 10.2/10.4 date format conversion.
  - Extracted some backtrace logging stuff from OFController into OFBacktrace
    for use on iOS.
  - Cleaned up iOS support in OFBundleRegistry and enable more on iOS.
  - Starting to remove the CDSA/CSSM references from OFXMLSignature's public
    API.
  - Add OFNetChangeNotifier, OFRelativeDateFormatter/Parser, and more
    Foundation extensions to the iOS build.
  - Fixed a crash encountered when calling -deepMutableCopy on an NSArray
    which contains an NSCFBoolean:  every object responds to the
    -mutableCopy selector (it's declared on NSObject), but not every
    object actually conforms to the NSMutableCopying protocol.
  - Fixed double archiving API on OFXML.

OmniFileStore:
  - Treat 406 result codes from PROPFIND the same as 404. This can happen
    when a DAV server has multiviews turned on and suggests an alternate.
    But since we are testing for the existance of a file, we don't want the
    alternate.
  - Added some UTI conveniences.
  - Use the right locale and time zone when parsing dates.
  - Fixed the WebDAV upload bug where choosing to rename the uploaded
    document (rather than replacing the existing document) would generate
    a bad HTTP request.

OmniAppKit:
  - Work around a lazy init bug in Cocoa scripting
  - Add NSAttributedString(OAExtensions) to iOS
  - Add locking for the global OAFontDescriptor uniquing table
  - Add tab stop support on OAParagraphStyle and hook up scalars for mutable
    paragraph styles.
  - Hook up paragraph styles in OATextStorage
  - Add OAFontDescriptorAttributeName, used in our other internal frameworks
    to record the originally requested font descriptor (independent of the
    closest font that was actually found).
  - A bit more implementation of OATextStorage

OmniQuartz:
  - Handle cornerRadius in CALayer vector output path.
  - Fall back to white for unarchiving HSV/pattern colors from XML on iOS
    instead of hitting OBFinishPorting.
  - Implement OQGetColorRefComponents() for iOS.
  - Added a utility routine for drawing SVG-style paths. This isn't complete
    (the translation of arcs is nontrivial) but it's potentially useful.

OmniUI/Text:
  - OUIEditableFrame now uses a text storage, somewhat similarly to the
    Cocoa text system. This allows the text frame to edit its mutable contents
    and respond to external mutations via -textStorageDidProcessEditing:.
  - Extended support for span background colors with an optional filter block
    to control which are drawn).
  - Extended OUITextLayout rendering to allow control for whether to render
    backgrounds, glyphs or both. This allows for proper layering of the
    selection color in cases that need it.
  - Added support for attempting to center the first line at a particular
    y-offset.
  - Added support for customizing inspector slices in text inspector.
  - Fixed bugs with rounding at non-integral scales.
  - Added support for losing first responder w/o ending editing. For example,
    the font size inspector has an editable field for setting the size.
  - Added more delegate hooks.
  - Added some support for simulated super/subscript since CoreText doesn't
    support it for arbitrary fonts.
  - Tons of other bug fixes and improvements.

OmniUI/Inspectors:
  - Inspectors are now fixed height, specified when they are created. The
    stack of slices is made scrollable if needed. Slices can still change
    height, but the popover will always be the same height (well, unless
    the keyboard forces it to be shorter).
  - Worked around various bugs with UIPopoverController.
  - Add support for a 'None' color type. Slices that allow 'None' also have
    a default color for when they are toggled away from None.
  - Simplify API on color slices to just return a single color; we never ended
    up needing an NSSet.
  - The inspected objects is now an NSArray instead of NSSet, allowing us to
    preserve the 'first' selected object (important when selecting text runs,
    for example).
  - When inspectors are updated, the get passed a reason that can allow them
    to perform less work. Typically they'll avoid reloading a UITableView
    since the calling code is going to do it in an animated way.
  - OUIActionInspectorSlices can be configured to start editing when they
    appear.
  - Color pickers selection is by identifier instead of index.
  - Can disable the automatic picking of the appropriate color pane for a
    given selection.
  - Split the font inspector into multiple slices and made them more
    subclassable for OmniOutliner.
  - Added a dark mode for OUIInspectorSegmentedControl.
  - Add support for placeholder text in OUIInspectorTextWell.
  - Better handling of memory warnings.
  - Improved support for automatically laying out inspector slices based on
    border edge insets.

OmniUI/Export:
  - Added support for importing/exporting file wrappers.
  - Improved support for picking an export format.

OmniUI/General:
  - Added OUICustomSubclass helpers to allow applications to specify subclasses
    to be used by superclasses that adopt its support.
  - We now do our own 'bar button item' popover presentation to avoid
    positioning bugs with custom view toolbar items.
  - Added a 'clear' toolbar button border type, used by OmniOutliner.
  - Lots of other smaller bug fixes.
  - Fixed up a bunch of document picker animation/performance issues.
  - Added a scroller on the docuemnt picker to quickly scroll through the list
    of documents.
  - More support for UTI mapping, exports, and email of nice attachments/HTML.
  - Added OUIGestureRecognizer superclass, for common behavior for some of our
    custom recognizers. Merged a bunch of work for OmniGraffle/iPad on long
    press and rotation recognizers.
  - Added support for displaying an image before the text in OUIOverlayView.
  - Added some debugging support for logging UIPopoverController size changes.

OmniUnzip:
  - Added conveniences for dealing with file wrappers, using OFFileWrapper
    for iOS support.

+++ Frameworks/OmniFoundation/DataStructures.subproj/OFEnumNameTable-OFXMLArchiving.m
@@ -1,4 +1,4 @@
-// Copyright 2002-2010 Omni Development, Inc.  All rights reserved.
+// Copyright 2002-2011 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -35,7 +35,7 @@ static int _compareIntptrs(const void *arg1, const void *arg2) {
 static inline int _writableEnumValue(NSInteger value)
 {
     // Foundation has started doing enum {...}; typedef NSUInteger FooEnum and the like.  We don't expect to get large values; let's assert that.
-    OBASSERT(value >= INT_MIN && value <= INT_MAX);
+    OBASSERT(value >= INT16_MIN && value <= INT16_MAX); // This assertion used to say INT_MIN and INT_MAX, but that assertion was failing on armv7 for values of 0 and 1.
     return (int)value;
 }
